Fix -Wsign-conversion warning in assert.t.cpp (issue #195)

+++ test/assert.t.cpp
@@ -80,7 +80,7 @@ int testAssume( int i, std::vector<int> const& v )
 {
     // This should compile without warnings.
     gsl_ASSUME( i >= 0 && static_cast<std::size_t>(i) < v.size() );
-    return v.at( i );
+    return v.at( static_cast<std::size_t>(i) );
 }
 
 // end of file
